Fix #32 : allow floats in sage exports configuration

+++ autonomie/compute/math_utils.py
@@ -82,7 +82,7 @@ def percentage(value, percent):
     """
         Return the value of the "percent" percent of the original "value"
     """
-    return int(float(value) * (int(percent)/100.0))
+    return int(float(value) * (float(percent)/100.0))
 
 
 def compute_taux(taux, value):
